commented out teleopOpenCV test - implemented OpenCV into auto
