+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/EasyOpenCVExample.java
@@ -1,164 +1,164 @@
-package org.firstinspires.ftc.teamcode;
-
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-
-import org.opencv.core.Core;
-import org.opencv.core.Mat;
-import org.opencv.core.Point;
-import org.opencv.core.Rect;
-import org.opencv.core.Scalar;
-import org.opencv.imgproc.Imgproc;
-import org.openftc.easyopencv.OpenCvCamera;
-import org.openftc.easyopencv.OpenCvCameraFactory;
-import org.openftc.easyopencv.OpenCvCameraRotation;
-import org.openftc.easyopencv.OpenCvInternalCamera;
-import org.openftc.easyopencv.OpenCvPipeline;
-import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
-
-
-@TeleOp(name="EasyOpenCVExample", group="UltimateGoal")
-public class EasyOpenCVExample extends LinearOpMode {
-
-    SkystoneDeterminationPipeline pipeline;
-
-    @Override
-    public void runOpMode()
-    {
-        WebcamName webcamName = hardwareMap.get(WebcamName.class, "webcam");
-        int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier("cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
-        final OpenCvCamera phoneCam = OpenCvCameraFactory.getInstance().createWebcam(webcamName, cameraMonitorViewId);
-        pipeline = new SkystoneDeterminationPipeline();
-        phoneCam.setPipeline(pipeline);
-
-        // We set the viewport policy to optimized view so the preview doesn't appear 90 deg
-        // out when the RC activity is in portrait. We do our actual image processing assuming
-        // landscape orientation, though.
-        phoneCam.setViewportRenderingPolicy(OpenCvCamera.ViewportRenderingPolicy.OPTIMIZE_VIEW);
-
-        phoneCam.openCameraDeviceAsync(new OpenCvCamera.AsyncCameraOpenListener()
-        {
-            @Override
-            public void onOpened()
-            {
-                phoneCam.startStreaming(320,240, OpenCvCameraRotation.SIDEWAYS_LEFT);
-            }
-        });
-
-        waitForStart();
-
-        while (opModeIsActive())
-        {
-            telemetry.addData("Analysis", pipeline.getAnalysis());
-            telemetry.addData("Position", pipeline.position);
-            telemetry.update();
-
-            // Don't burn CPU cycles busy-looping in this sample
-            sleep(50);
-        }
-    }
-
-    public static class SkystoneDeterminationPipeline extends OpenCvPipeline
-    {
-        /*
-         * An enum to define the skystone position
-         */
-        public enum RingPosition
-        {
-            FOUR,
-            ONE,
-            NONE
-        }
-
-        /*
-         * Some color constants
-         */
-        static final Scalar BLUE = new Scalar(0, 0, 255);
-        static final Scalar GREEN = new Scalar(0, 255, 0);
-
-        /*
-         * The core values which define the location and size of the sample regions
-         */
-        static final Point REGION1_TOPLEFT_ANCHOR_POINT = new Point(181,98);
-
-        static final int REGION_WIDTH = 35;
-        static final int REGION_HEIGHT = 25;
-
-        final int FOUR_RING_THRESHOLD = 150;
-        final int ONE_RING_THRESHOLD = 135;
-
-        Point region1_pointA = new Point(
-                REGION1_TOPLEFT_ANCHOR_POINT.x,
-                REGION1_TOPLEFT_ANCHOR_POINT.y);
-        Point region1_pointB = new Point(
-                REGION1_TOPLEFT_ANCHOR_POINT.x + REGION_WIDTH,
-                REGION1_TOPLEFT_ANCHOR_POINT.y + REGION_HEIGHT);
-
-        /*
-         * Working variables
-         */
-        Mat region1_Cb;
-        Mat YCrCb = new Mat();
-        Mat Cb = new Mat();
-        int avg1;
-
-        // Volatile since accessed by OpMode thread w/o synchronization
-        private volatile RingPosition position = RingPosition.FOUR;
-
-        /*
-         * This function takes the RGB frame, converts to YCrCb,
-         * and extracts the Cb channel to the 'Cb' variable
-         */
-        void inputToCb(Mat input)
-        {
-            Imgproc.cvtColor(input, YCrCb, Imgproc.COLOR_RGB2YCrCb);
-            Core.extractChannel(YCrCb, Cb, 1);
-        }
-
-        @Override
-        public void init(Mat firstFrame)
-        {
-            inputToCb(firstFrame);
-
-            region1_Cb = Cb.submat(new Rect(region1_pointA, region1_pointB));
-        }
-
-        @Override
-        public Mat processFrame(Mat input)
-        {
-            inputToCb(input);
-
-            avg1 = (int) Core.mean(region1_Cb).val[0];
-
-            Imgproc.rectangle(
-                    input, // Buffer to draw on
-                    region1_pointA, // First point which defines the rectangle
-                    region1_pointB, // Second point which defines the rectangle
-                    BLUE, // The color the rectangle is drawn in
-                    2); // Thickness of the rectangle lines
-
-            position = RingPosition.FOUR; // Record our analysis
-            if(avg1 > FOUR_RING_THRESHOLD){
-                position = RingPosition.FOUR;
-            }else if (avg1 > ONE_RING_THRESHOLD){
-                position = RingPosition.ONE;
-            }else{
-                position = RingPosition.NONE;
-            }
-
-            Imgproc.rectangle(
-                    input, // Buffer to draw on
-                    region1_pointA, // First point which defines the rectangle
-                    region1_pointB, // Second point which defines the rectangle
-                    GREEN, // The color the rectangle is drawn in
-                    -1); // Negative thickness means solid fill
-
-            return input;
-        }
-
-        public int getAnalysis()
-        {
-            return avg1;
-        }
-    }
-}
+//package org.firstinspires.ftc.teamcode;
+//
+//import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
+//import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+//
+//import org.opencv.core.Core;
+//import org.opencv.core.Mat;
+//import org.opencv.core.Point;
+//import org.opencv.core.Rect;
+//import org.opencv.core.Scalar;
+//import org.opencv.imgproc.Imgproc;
+//import org.openftc.easyopencv.OpenCvCamera;
+//import org.openftc.easyopencv.OpenCvCameraFactory;
+//import org.openftc.easyopencv.OpenCvCameraRotation;
+//import org.openftc.easyopencv.OpenCvInternalCamera;
+//import org.openftc.easyopencv.OpenCvPipeline;
+//import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
+//
+//
+//@TeleOp(name="EasyOpenCVExample", group="UltimateGoal")
+//public class EasyOpenCVExample extends LinearOpMode {
+//
+//    SkystoneDeterminationPipeline pipeline;
+//
+//    @Override
+//    public void runOpMode()
+//    {
+//        WebcamName webcamName = hardwareMap.get(WebcamName.class, "webcam");
+//        int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier("cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
+//        final OpenCvCamera phoneCam = OpenCvCameraFactory.getInstance().createWebcam(webcamName, cameraMonitorViewId);
+//        pipeline = new SkystoneDeterminationPipeline();
+//        phoneCam.setPipeline(pipeline);
+//
+//        // We set the viewport policy to optimized view so the preview doesn't appear 90 deg
+//        // out when the RC activity is in portrait. We do our actual image processing assuming
+//        // landscape orientation, though.
+//        phoneCam.setViewportRenderingPolicy(OpenCvCamera.ViewportRenderingPolicy.OPTIMIZE_VIEW);
+//
+//        phoneCam.openCameraDeviceAsync(new OpenCvCamera.AsyncCameraOpenListener()
+//        {
+//            @Override
+//            public void onOpened()
+//            {
+//                phoneCam.startStreaming(320,240, OpenCvCameraRotation.SIDEWAYS_LEFT);
+//            }
+//        });
+//
+//        waitForStart();
+//
+//        while (opModeIsActive())
+//        {
+//            telemetry.addData("Analysis", pipeline.getAnalysis());
+//            telemetry.addData("Position", pipeline.position);
+//            telemetry.update();
+//
+//            // Don't burn CPU cycles busy-looping in this sample
+//            sleep(50);
+//        }
+//    }
+//
+//    public static class SkystoneDeterminationPipeline extends OpenCvPipeline
+//    {
+//        /*
+//         * An enum to define the skystone position
+//         */
+//        public enum RingPosition
+//        {
+//            FOUR,
+//            ONE,
+//            NONE
+//        }
+//
+//        /*
+//         * Some color constants
+//         */
+//        static final Scalar BLUE = new Scalar(0, 0, 255);
+//        static final Scalar GREEN = new Scalar(0, 255, 0);
+//
+//        /*
+//         * The core values which define the location and size of the sample regions
+//         */
+//        static final Point REGION1_TOPLEFT_ANCHOR_POINT = new Point(181,98);
+//
+//        static final int REGION_WIDTH = 35;
+//        static final int REGION_HEIGHT = 25;
+//
+//        final int FOUR_RING_THRESHOLD = 150;
+//        final int ONE_RING_THRESHOLD = 135;
+//
+//        Point region1_pointA = new Point(
+//                REGION1_TOPLEFT_ANCHOR_POINT.x,
+//                REGION1_TOPLEFT_ANCHOR_POINT.y);
+//        Point region1_pointB = new Point(
+//                REGION1_TOPLEFT_ANCHOR_POINT.x + REGION_WIDTH,
+//                REGION1_TOPLEFT_ANCHOR_POINT.y + REGION_HEIGHT);
+//
+//        /*
+//         * Working variables
+//         */
+//        Mat region1_Cb;
+//        Mat YCrCb = new Mat();
+//        Mat Cb = new Mat();
+//        int avg1;
+//
+//        // Volatile since accessed by OpMode thread w/o synchronization
+//        private volatile RingPosition position = RingPosition.FOUR;
+//
+//        /*
+//         * This function takes the RGB frame, converts to YCrCb,
+//         * and extracts the Cb channel to the 'Cb' variable
+//         */
+//        void inputToCb(Mat input)
+//        {
+//            Imgproc.cvtColor(input, YCrCb, Imgproc.COLOR_RGB2YCrCb);
+//            Core.extractChannel(YCrCb, Cb, 1);
+//        }
+//
+//        @Override
+//        public void init(Mat firstFrame)
+//        {
+//            inputToCb(firstFrame);
+//
+//            region1_Cb = Cb.submat(new Rect(region1_pointA, region1_pointB));
+//        }
+//
+//        @Override
+//        public Mat processFrame(Mat input)
+//        {
+//            inputToCb(input);
+//
+//            avg1 = (int) Core.mean(region1_Cb).val[0];
+//
+//            Imgproc.rectangle(
+//                    input, // Buffer to draw on
+//                    region1_pointA, // First point which defines the rectangle
+//                    region1_pointB, // Second point which defines the rectangle
+//                    BLUE, // The color the rectangle is drawn in
+//                    2); // Thickness of the rectangle lines
+//
+//            position = RingPosition.FOUR; // Record our analysis
+//            if(avg1 > FOUR_RING_THRESHOLD){
+//                position = RingPosition.FOUR;
+//            }else if (avg1 > ONE_RING_THRESHOLD){
+//                position = RingPosition.ONE;
+//            }else{
+//                position = RingPosition.NONE;
+//            }
+//
+//            Imgproc.rectangle(
+//                    input, // Buffer to draw on
+//                    region1_pointA, // First point which defines the rectangle
+//                    region1_pointB, // Second point which defines the rectangle
+//                    GREEN, // The color the rectangle is drawn in
+//                    -1); // Negative thickness means solid fill
+//
+//            return input;
+//        }
+//
+//        public int getAnalysis()
+//        {
+//            return avg1;
+//        }
+//    }
+//}
